backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
transformers
qdrant-client
google-generativeai
sqlalchemy
psycopg2-binary
python-dotenv
